Remove v18 from bitcoind-rpc
