+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/common/BitcoindRpcClient.scala
@@ -22,7 +22,6 @@ import org.bitcoins.crypto.{
   DoubleSha256DigestBE,
   StringFactory
 }
-import org.bitcoins.rpc.client.v18.BitcoindV18RpcClient
 import org.bitcoins.rpc.client.v19.BitcoindV19RpcClient
 import org.bitcoins.rpc.client.v20.BitcoindV20RpcClient
 import org.bitcoins.rpc.client.v21.BitcoindV21RpcClient
@@ -344,7 +343,6 @@ object BitcoindRpcClient {
   def fromVersion(version: BitcoindVersion, instance: BitcoindInstance)(implicit
       system: ActorSystem): BitcoindRpcClient = {
     val bitcoind = version match {
-      case BitcoindVersion.V18 => BitcoindV18RpcClient.withActorSystem(instance)
       case BitcoindVersion.V19 => BitcoindV19RpcClient.withActorSystem(instance)
       case BitcoindVersion.V20 => BitcoindV20RpcClient.withActorSystem(instance)
       case BitcoindVersion.V21 => BitcoindV21RpcClient.withActorSystem(instance)
@@ -373,13 +371,9 @@ object BitcoindVersion extends StringFactory[BitcoindVersion] with Logging {
   val newest: BitcoindVersion = V23
 
   val standard: Vector[BitcoindVersion] =
-    Vector(V18, V19, V20, V21, V22, V23)
+    Vector(V19, V20, V21, V22, V23)
 
   val known: Vector[BitcoindVersion] = standard
-
-  case object V18 extends BitcoindVersion {
-    override def toString: String = "v0.18"
-  }
 
   case object V19 extends BitcoindVersion {
     override def toString: String = "v0.19"
@@ -419,7 +413,6 @@ object BitcoindVersion extends StringFactory[BitcoindVersion] with Logging {
   def fromNetworkVersion(int: Int): BitcoindVersion = {
     //need to translate the int 210100 (as an example) to a BitcoindVersion
     int.toString.substring(0, 2) match {
-      case "18" => V18
       case "19" => V19
       case "20" => V20
       case "21" => V21
