+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/common/BlockchainRpc.scala
@@ -29,8 +29,6 @@ trait BlockchainRpc { self: Client =>
 
   def getBlockChainInfo: Future[GetBlockChainInfoResult] = {
     self.version.flatMap {
-      case V18 =>
-        bitcoindCall[GetBlockChainInfoResultPreV19]("getblockchaininfo")
       case V22 | V21 | V20 | V19 =>
         bitcoindCall[GetBlockChainInfoResultPostV19]("getblockchaininfo")
       case V23 | Unknown =>
@@ -85,7 +83,7 @@ trait BlockchainRpc { self: Client =>
           "getblock",
           List(JsString(headerHash.hex), isVerboseJsonObject))
 
-      case V18 | V19 | V20 | V21 =>
+      case V19 | V20 | V21 =>
         bitcoindCall[GetBlockWithTransactionsResultPreV22](
           "getblock",
           List(JsString(headerHash.hex), isVerboseJsonObject))
