+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/common/MultisigRpc.scala
@@ -46,7 +46,7 @@ trait MultisigRpc { self: Client =>
           "addmultisigaddress",
           params,
           uriExtensionOpt = walletNameOpt.map(walletExtension))
-      case V18 | V19 =>
+      case V19 =>
         bitcoindCall[MultiSigResultPreV20]("addmultisigaddress",
                                            params,
                                            uriExtensionOpt =
@@ -92,7 +92,7 @@ trait MultisigRpc { self: Client =>
                Json.toJson(addressType)),
           uriExtensionOpt = walletNameOpt.map(walletExtension)
         )
-      case V18 | V19 =>
+      case V19 =>
         bitcoindCall[MultiSigResultPreV20](
           "createmultisig",
           List(JsNumber(minSignatures), Json.toJson(keys.map(_.hex))),
