+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/common/TransactionRpc.scala
@@ -8,7 +8,6 @@ import org.bitcoins.core.protocol.blockchain.MerkleBlock
 import org.bitcoins.crypto.{DoubleSha256Digest, DoubleSha256DigestBE}
 import org.bitcoins.rpc.client.common.BitcoindVersion.{
   Unknown,
-  V18,
   V19,
   V20,
   V21,
@@ -94,7 +93,7 @@ trait TransactionRpc { self: Client =>
           "gettxout",
           List(JsString(txid.hex), JsNumber(vout), JsBoolean(includeMemPool)))
 
-      case V18 | V19 | V20 | V21 =>
+      case V19 | V20 | V21 =>
         bitcoindCall[GetTxOutResultPreV22](
           "gettxout",
           List(JsString(txid.hex), JsNumber(vout), JsBoolean(includeMemPool)))
@@ -112,7 +111,7 @@ trait TransactionRpc { self: Client =>
             "gettxout",
             List(JsString(txid.hex), JsNumber(vout), JsBoolean(includeMemPool)))
 
-        case V18 | V19 | V20 | V21 =>
+        case V19 | V20 | V21 =>
           bitcoindCall[GetTxOutResultPreV22](
             "gettxout",
             List(JsString(txid.hex), JsNumber(vout), JsBoolean(includeMemPool)))
