+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/config/BitcoindInstance.scala
@@ -55,8 +55,6 @@ sealed trait BitcoindInstanceLocal extends BitcoindInstance {
         .last
 
     foundVersion match {
-      case _: String if foundVersion.startsWith(BitcoindVersion.V18.toString) =>
-        BitcoindVersion.V18
       case _: String if foundVersion.startsWith(BitcoindVersion.V19.toString) =>
         BitcoindVersion.V19
       case _: String if foundVersion.startsWith(BitcoindVersion.V20.toString) =>
